% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container_exec.R
\name{container_exec}
\alias{container_exec}
\title{Title
container_exec
Description
run commands inside the container}
\usage{
container_exec(containerid, command, use_podman = FALSE)
}
\arguments{
\item{containerid}{the name of the container}

\item{command}{commands need to be run inside the container}

\item{use_podman}{logit, use podman as  backend when it is TRUE}
}
\value{
str
}
\description{
Title
container_exec
Description
run commands inside the container
}
\examples{
container_exec("rsshubnew","pwd")
}
