% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container_run.R
\name{container_run}
\alias{container_run}
\title{Title
container_run
Description
run container at the first time}
\usage{
container_run(
  imageid,
  name,
  volume = NULL,
  ports,
  environmentvarible = NULL,
  ifD = TRUE,
  restart = TRUE,
  run_at_once = TRUE,
  use_podman = FALSE
)
}
\arguments{
\item{imageid}{the image label}

\item{name}{the name of the container}

\item{volume}{the volume setting of the container}

\item{ports}{the ports setting of the container}

\item{environmentvarible}{the "-e" options of docker}

\item{ifD}{the "-d" option of docker}

\item{restart}{the "--restart=always" option}

\item{run_at_once}{whether using "create" or "run"}

\item{use_podman}{logit, use podman as  backend when it is TRUE}
}
\value{
str
}
\description{
Title
container_run
Description
run container at the first time
}
\examples{
container_run("diygod/rsshub","rsshubnew",ports = "-p 1300:1200")
}
