% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container_update.R
\name{container_update}
\alias{container_update}
\title{Title
container_update
Description
update the container when the upstream image changes}
\usage{
container_update(
  containerid,
  imageid,
  force_rm_container = FALSE,
  volume = NULL,
  ports,
  environmentvarible = NULL,
  ifD = TRUE,
  restart = TRUE,
  run_at_once = TRUE,
  use_podman = FALSE
)
}
\arguments{
\item{containerid}{the name of the container}

\item{imageid}{the image label}

\item{force_rm_container}{force remove the container if its true}

\item{volume}{the volume setting of the container}

\item{ports}{the ports setting of the container}

\item{environmentvarible}{the "-e" options of docker}

\item{ifD}{the "-d" option of docker}

\item{restart}{the "--restart=always" option}

\item{run_at_once}{whether using "create" or "run"}

\item{use_podman}{logit, use podman as  backend when it is TRUE}
}
\value{
str
}
\description{
Title
container_update
Description
update the container when the upstream image changes
}
\examples{
container_update("rsshubnew","diygod/rsshub",ports = "-p 1300:1200",restart = FALSE)
}
