% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_stats.R
\name{docker_stats}
\alias{docker_stats}
\title{container stats}
\usage{
docker_stats(base_command = "docker stats --no-stream", additional = NA)
}
\arguments{
\item{base_command}{the base command}

\item{additional}{additional command}
}
\value{
data.frame
}
\description{
show the container stats in a data.frame
}
\examples{
docker_stats()
}
