% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_build.R
\name{image_build}
\alias{image_build}
\title{Title
image_build
Description
build docker image from Dockerfile}
\usage{
image_build(name, folder_dockerfile)
}
\arguments{
\item{name}{the label of the built image}

\item{folder_dockerfile}{the dir where the dockerfile lies in}
}
\value{
str
}
\description{
Title
image_build
Description
build docker image from Dockerfile
}
\examples{
image_build("fallingstar10/gmadeomicsrocker:base","./dev/gmade_rocker-base")
}
