% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_load.R
\name{image_load}
\alias{image_load}
\title{Title
image_load
Description
load tarfile to docker image}
\usage{
image_load(tarfile, import = FALSE, imageid = NULL)
}
\arguments{
\item{tarfile}{the dest output tarfile,with a full dir and its name}

\item{import}{not using "load" command when its true}

\item{imageid}{the image label}
}
\value{
str
}
\description{
Title
image_load
Description
load tarfile to docker image
}
\examples{
image_load(tarfile = "./dev/rsshub.tar",import = TRUE,imageid = "rsshub:load")
}
