% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_push.R
\name{image_push}
\alias{image_push}
\title{Title
image_push
Description
push image to dockerhub}
\usage{
image_push(
  containername,
  username = "example",
  password = "example",
  auther = FALSE
)
}
\arguments{
\item{containername}{the label of the pushed image}

\item{username}{the username of dockerhub account}

\item{password}{the password of dockerhub account}

\item{auther}{if the a dockerhub account has previously login,then auther should be FALSE}
}
\value{
str
}
\description{
Title
image_push
Description
push image to dockerhub
}
\examples{
image_push("fallingstar10/gmadeomicsrocker:base",username = "fallingstar10")
}
