% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_save.R
\name{image_save}
\alias{image_save}
\title{Title
image_save
Description
save image as tarfile}
\usage{
image_save(imageid, tarfile, use_podman = FALSE)
}
\arguments{
\item{imageid}{the image label}

\item{tarfile}{the dest output tarfile,with a full dir and its name}

\item{use_podman}{logit, use podman as  backend when it is TRUE}
}
\value{
str
}
\description{
Title
image_save
Description
save image as tarfile
}
\examples{
image_save("diygod/rsshub","./rsshub.tar")
}
