% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_tag.R
\name{image_tag}
\alias{image_tag}
\title{Title
image_tag
Description
change the tag of image}
\usage{
image_tag(oldtag, newtag, .keep = TRUE, use_podman = FALSE)
}
\arguments{
\item{oldtag}{the old/original image label,str}

\item{newtag}{the new image label,str}

\item{.keep}{whether to keep the oldtag image}

\item{use_podman}{logit, use podman as  backend when it is TRUE}
}
\value{
str
}
\description{
Title
image_tag
Description
change the tag of image
}
\examples{
  image_tag("diygod/rsshub","fall/rsshub")
}
