% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerr.R
\name{inspect}
\alias{inspect}
\title{list active docker containers}
\usage{
inspect(container)
}
\arguments{
\item{container}{the container you want to inspect}
}
\value{
a tibble output of the docker container ls command
}
\description{
list active docker containers
}
