% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsimage.R
\name{lsimage}
\alias{lsimage}
\title{Title
lsimage
Description
list all images}
\usage{
lsimage(base_command = "docker images", additional = NA, use_podman = FALSE)
}
\arguments{
\item{base_command}{the base command}

\item{additional}{additional command}

\item{use_podman}{logit, use podman as  backend when it is TRUE}
}
\value{
data.frame
}
\description{
Title
lsimage
Description
list all images
}
\examples{
lsimage()
}
