% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerr.R
\name{terminal2tbl}
\alias{terminal2tbl}
\title{transform terminal output to a tibble}
\usage{
terminal2tbl(stdout)
}
\arguments{
\item{stdout}{output from terminal}
}
\description{
transform terminal output to a tibble
}
