% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container_control.R
\name{container_control}
\alias{container_control}
\title{Title
container_control
Description
control the statue of containers by stop, start ,restart and delete(rm)}
\usage{
container_control(
  containerid,
  mod = c("stop", "start", "restart", "rm"),
  forcerm = FALSE
)
}
\arguments{
\item{containerid}{the label of container}

\item{mod}{the control mod,including stop, start ,restart and delete(rm)}

\item{forcerm}{force remove the container if its true}
}
\value{
str
}
\description{
Title
container_control
Description
control the statue of containers by stop, start ,restart and delete(rm)
}
\examples{
container_control("rsshubnew","restart")
}
