% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container_dockercompose.R
\name{container_dockercompose}
\alias{container_dockercompose}
\title{Title
container_dockercompose
Description
initial container by docker compose}
\usage{
container_dockercompose(yamlfile, use_podman = FALSE)
}
\arguments{
\item{yamlfile}{the docker compose yaml file}

\item{use_podman}{logit, use podman as  backend when it is TRUE}
}
\value{
str
}
\description{
Title
container_dockercompose
Description
initial container by docker compose
}
\examples{
container_dockercompose("./dev/nextcloud_compose.yaml")
}
