% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_version.R
\name{docker_version}
\alias{docker_version}
\title{docker information}
\usage{
docker_version()
}
\value{
str
}
\description{
display the information of the installed docker
}
\examples{
docker_version()
}
