% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_pull.R
\name{image_pull}
\alias{image_pull}
\title{Title
image_pull
Description
pull image from dockerhub}
\usage{
image_pull(id, use_podman = FALSE)
}
\arguments{
\item{id}{the image id ,str}

\item{use_podman}{logit, use podman as  backend when it is TRUE}
}
\value{
str
}
\description{
Title
image_pull
Description
pull image from dockerhub
}
\examples{
image_pull("frooodle/s-pdf:latest")
}
