% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lscontainer.R
\name{lscontainer}
\alias{lscontainer}
\title{Add one to any value}
\usage{
lscontainer(base_command = "docker ps", additional = "-a", use_podman = FALSE)
}
\arguments{
\item{base_command}{A str}

\item{additional}{A str}

\item{use_podman}{logit, use podman as  backend when it is TRUE}
}
\value{
data.frame
}
\description{
Add one to any value
}
\examples{
lscontainer(additional = "-a")
}
